Word2Vec_Sample/requirements.txt
dotenv
mysql-connector-python
google-api-python-client
pandas
konlpy
gensim
python-mecab-ko
flask